refactor: reorganize files into lib
